os/bluestore: fix huge (>4GB) bluefs reads
(cherry picked from commit fe558c4dd12b25b2d523987f22fde5b5898cfb83)

 Conflicts:
(still present buf param in BlueFS::read, lack of )https://github.com/ceph/ceph/pull/34421)
	src/os/bluestore/BlueFS.h
	src/os/bluestore/BlueRocksEnv.cc
	src/test/objectstore/test_bluefs.cc

+++ src/test/objectstore/test_bluefs.cc
@@ -304,7 +304,7 @@ TEST(BlueFS, very_large_write) {
     unique_ptr<char> huge_buf(new char[h->file->fnode.size]);
     auto l = h->file->fnode.size;
     int64_t r = fs.read(h, &readbuf, 0, l, NULL, huge_buf.get());
-    ASSERT_EQ(r, l);
+    ASSERT_EQ(r, (int64_t)l);
     delete h;
   }
   fs.umount();
